+++ apps/ungdomsytelse-deltaker/vite.shared.config.ts
@@ -2,7 +2,7 @@
 import react from '@vitejs/plugin-react';
 import checker from 'vite-plugin-checker';
 import tailwindcss from '@tailwindcss/vite';
-import { getDevAppSettings } from './src/devAppSettings';
+import { getDevAppSettings } from './mock/devAppSettings';
 
 // Felles plugins som brukes i alle konfigurasjoner
 export const commonPlugins = [