+++ apps/ungdomsytelse-veileder/vite.dev.config.ts
@@ -11,6 +11,8 @@ const appSettings: AppEnv = {
     PUBLIC_PATH: '/sif-brukerdialog/ungdomsytelse-veileder',
     GITHUB_REF_NAME: 'local',
 
+    SIF_PUBLIC_FEATURE_ENDRE_SLUTTDATO: 'on',
+
     UNG_DELTAKELSE_OPPLYSER_FRONTEND_PATH: '/api/ung-deltakelse-opplyser',
     UNG_DELTAKELSE_OPPLYSER_API_URL: 'http://localhost:8089',
     UNG_DELTAKELSE_OPPLYSER_API_SCOPE: 'dev-gcp:dusseldorf:ung-deltakelse-opplyser',