+++ packages/appstatus-react-ds/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @navikt/appstatus-react-ds
 
+## 2.20.30
+
+### Patch Changes
+
+- Pakkeoppdateringer - minor/patch
+
+## 2.20.29
+
+### Patch Changes
+
+- Oppdaterte pakker - minor/patch + codegen
+
 ## 2.20.28
 
 ### Patch Changes