+++ packages/k9-brukerdialog-prosessering-api/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @navikt/k9-brukerdialog-prosessering-api
 
+## 0.1.10
+
+### Patch Changes
+
+- Pakkeoppdateringer - minor/patch
+
+## 0.1.9
+
+### Patch Changes
+
+- Oppdaterte pakker - minor/patch + codegen
+
 ## 0.1.8
 
 ### Patch Changes